Fixed bug which could cause excessively slow computation of low point densities - far more points were being considered than necessary due to a wraparound doing unsigned arithmetic when negative results were possible. Tested with the code below, showing that performance remains the same for computing large densities but is vastly improved for computing small densities.
import os

import IECore

os.environ["CORTEX_POINTDISTRIBUTION_TILESET"] = "test/IECore/data/pointDistributions/pointDistributionTileSet2048.dat"
pd = IECore.PointDistribution.defaultInstance()

bound = IECore.Box2f( IECore.V2f( 0 ), IECore.V2f( 1 ) )

t = IECore.Timer()

points = pd( bound, 20000000 )

print t.stop()

size = IECore.V2f( 0.05 )
t.start()

for i in range( 0, 1000 ) :
	
	min = IECore.Rand32().nextV2f()
	points = pd( IECore.Box2f( min, min + size ), 20000000 )

print t.stop()

size = IECore.V2f( 0.05 )
t.start()

for i in range( 0, 10000 ) :
	
	min = IECore.Rand32().nextV2f()
	points = pd( IECore.Box2f( min, min + size ), 1000 )

print t.stop()

# results
#########################################
# baseline
#
#	0.61	3.08	3.48
#	0.61	3.08	3.48
#	0.59	3.09	3.48
#
# with fixed unsigned underflow
#
#	0.61	3.08	0.22
#	0.61	3.08	0.22
#	0.62	3.08	0.21

+++ include/IECore/PointDistribution.inl
@@ -127,11 +127,11 @@ void PointDistribution::recurseTile( const Tile &tile, const Imath::V2f &bottomL
 	
 	float tileArea = tileSize * tileSize;
 	float numPointsInTile = density * tileArea;
-	unsigned potentialPoints = std::min( tile.subPoints.size(), (size_t)numPointsInTile - tile.points.size() );
+	int potentialPoints = std::min( (int)tile.subPoints.size(), (int)numPointsInTile - (int)tile.points.size() );
 	
 	float factor = 1.0f / ( numPointsInTile );
 	
-	for( unsigned i=0; i<potentialPoints; i++ )
+	for( int i=0; i<potentialPoints; i++ )
 	{
 		const Imath::V2f &p = bottomLeft + tile.subPoints[i] * tileSize;
 		if( !bounds.intersects( p ) )
